put the I2C-specific correction writer in the I2C-specific file
