+++ components/movementsensor/gpsrtkserial/pmtk.go
@@ -5,6 +5,8 @@ package gpsrtk
 
 import (
 	"context"
+	"errors"
+	"io"
 
 	"go.viam.com/rdk/components/board/genericlinux/buses"
 	"go.viam.com/rdk/components/movementsensor"
@@ -134,7 +136,7 @@ func makeRTKI2C(
 	}
 	g.cachedData = gpsutils.NewCachedData(dev, logger)
 
-	g.correctionWriter, err = NewCorrectionWriter(newConf.I2CBus, byte(newConf.I2CAddr))
+	g.correctionWriter, err = newCorrectionWriter(newConf.I2CBus, byte(newConf.I2CAddr))
 	if err != nil {
 		return nil, err
 	}
@@ -143,4 +145,42 @@ func makeRTKI2C(
 		return nil, err
 	}
 	return g, nil
+}
+
+func newCorrectionWriter(busname string, address byte) (io.ReadWriteCloser, error) {
+	bus, err := buses.NewI2cBus(busname)
+	if err != nil {
+		return nil, err
+	}
+	handle, err := bus.OpenHandle(address)
+	if err != nil {
+		return nil, err
+	}
+	correctionWriter := i2cCorrectionWriter {
+		bus: bus,
+		handle: handle,
+	}
+	return &correctionWriter, nil
+}
+
+// This implements the io.ReadWriteCloser interface.
+type i2cCorrectionWriter struct {
+	bus    buses.I2C
+	handle buses.I2CHandle
+}
+
+func (i *i2cCorrectionWriter) Read(p []byte) (int, error) {
+	return 0, errors.New("unimplemented")
+}
+
+func (i *i2cCorrectionWriter) Write(p []byte) (int, error) {
+	err := i.handle.Write(context.Background(), p)
+	if err != nil {
+		return 0, err
+	}
+	return len(p), nil
+}
+
+func (i *i2cCorrectionWriter) Close() error {
+	return i.handle.Close()
 }
